+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ are used for versioning (schema follows below):
 2019-xx-xx (not-yet-released)
 
 - Prevent unicode errors in tests on Python 2.7.
+- Fixes in occasionally failing search test (`test_search`).
 
 0.17.6
 ------